refactor(runtime-operator): extract PrecompileReconciler into production package
